Remove CA1D. Ploting not working expectedly
