+++ src/Agents.jl
@@ -20,7 +20,4 @@ include("simulations/collect.jl")
 include("simulations/paramscan.jl")
 include("simulations/sample.jl")
 
-# Auxilary
-include("CA1D.jl")
-
 end # module
